Update proposal.tex
I have added your completed section to the Overleaf document, including the abstract. I finished the subsubsection on "The Rust Standard Library" as well as the sections on Verus and Rust’s Affine Type System, based on the three points you provided. Additionally, I appended my content for the following sections and made some minor revisions to your text for grammatical accuracy.

One issue I noticed is that there doesn't appear to be a reference list at the end.